rename matchmaker.ts to client.ts
